SKUAnomalyDetector/requirements.txt
torch>=2.0.0
transformers>=4.36.0
scikit-learn>=1.3.0
pillow>=9.0.0
numpy>=1.23.0
pandas>=1.5.0
